rename: model -> domain 패키지명 변경
