+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,6 +1,5 @@
-package roomescape.model;
+package roomescape.domain;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
