+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.model.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationResponse(
         Long id,
